fix Spear return
Spear 돌아오는 속도 state Manager에서 받아옴
