+++ Assets/Prefabs/JSW/Spear.prefab
@@ -259,6 +259,7 @@ MonoBehaviour:
   returnSpeed: 3
   targetPosition: {x: 0, y: 0, z: 0}
   startPosition: {x: 0, y: 0, z: 0}
+  tail: {fileID: 4149198655838722403}
   acceleration: 2
 --- !u!1 &7605711517021760476
 GameObject:
