--- Assets/Script/JSW/Weapon/Spear1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0eba44ec39e0c494ea841f56fa590b5d
